+++ Engine/Ball.h
@@ -7,16 +7,16 @@ public:
 	Ball();
 	int Update();
 	void RandDirection();
-	int GetX(){ return x; }
-	int GetY(){ return y; }
-	void Shot(int XV);
+	int GetX(){ return static_cast<int>(x); }
+	int GetY(){ return static_cast<int>(y); }
+	void Shot(float XV);
 	void Collision(Paddle *paddle);
 	static constexpr int height = 20;
 	static constexpr int width = 20;
 
 private:
-	int x;
-	int y;
-	int xv;
-	int yv;
+	float x;
+	float y;
+	float xv;
+	float yv;
 };
